Add RenderPaletted and greatly improve performance
RenderPaletted accepts 3-color or full color images and automatically
converts and renders them to the display. See documentation for deatils.

This changes the rendering time on a Raspberry Pi 3b from 3.5s to
~250ms. The biggest savings came from batching data over the wire (2k
from 1 byte), and increasing the clock rate from 5MHz to 20MHz, which is
the maximum specified in the manual.

Other savings came from implementation changes.

Dropped Convert API.

Changed Render to accept byte slices. Also, the second image in render
is now expected to be inverted already, which helps us send data faster.